+++ packages/autometrics/src/index.ts
@@ -1,6 +1,5 @@
 export * from "./buildInfo";
-export { registerExporter } from "./exporter";
-export * from "./histograms";
+export { registerExporter } from "./instrumentation";
 export * as amLogger from "./logger";
 export * from "./objectives";
 export * from "./wrappers";
